requirements.txt
asgiref==3.6.0
bootstrap4==0.1.0
certifi==2022.12.7
charset-normalizer==3.1.0
crispy-bootstrap4==2022.1
crispy-bootstrap5==0.7
DateTime==5.1
dj-database-url==1.3.0
Django==3.2.18
django-ckeditor==6.3.2
django-crispy-forms==2.0
django-heroku==0.3.1
django-js-asset==2.0.0
gunicorn==20.1.0
idna==3.4
numpy==1.21.6
pandas==1.1.5
psycopg2==2.9.6
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3
requests==2.28.2
six==1.16.0
sqlparse==0.4.3
typing_extensions==4.5.0
urllib3==1.26.15
whitenoise==6.4.0
zope.interface==6.0
